backend/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
dateutil>=2.8.2
python-dateutil>=2.8.2
google-generativeai>=0.3.0
supabase-py>=2.0.0
# For TTS (if using gTTS or similar, add here)
# gTTS>=2.3.2
